% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_releases.R
\name{download_release_data}
\alias{download_release_data}
\title{Download data from GitHub Release}
\usage{
download_release_data(repo, dest_dir = "data", tag = "data")
}
\arguments{
\item{repo}{Repository in "owner/repo" format}

\item{dest_dir}{Destination directory for downloaded files}

\item{tag}{Release tag (default: "data")}
}
\value{
Invisible path to destination directory
}
\description{
Downloads all data files from the specified release tag.
}
\examples{
\dontrun{
download_release_data("peteowen1/kworb", "data")
}
}
