% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_chart.R
\name{get_country_codes}
\alias{get_country_codes}
\title{Get available country codes}
\usage{
get_country_codes()
}
\value{
Character vector of country codes
}
\description{
Returns a character vector of known country codes that can be used
with \code{scrape_chart()}.
}
\examples{
codes <- get_country_codes()
head(codes)
}
