% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{get_storage_summary}
\alias{get_storage_summary}
\title{Get storage summary}
\usage{
get_storage_summary(base_path = "data")
}
\arguments{
\item{base_path}{Base data directory}
}
\value{
List with storage statistics
}
\description{
Returns summary statistics about stored data.
}
