% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_releases.R
\name{list_release_files}
\alias{list_release_files}
\title{List files in GitHub Release}
\usage{
list_release_files(repo, tag = "data")
}
\arguments{
\item{repo}{Repository in "owner/repo" format}

\item{tag}{Release tag (default: "data")}
}
\value{
tibble with file information
}
\description{
Returns information about files stored in a release.
}
