% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{list_stored_tracks}
\alias{list_stored_tracks}
\title{List available track IDs in storage}
\usage{
list_stored_tracks(path = "data/tracks")
}
\arguments{
\item{path}{Directory path containing track CSVs}
}
\value{
Character vector of track IDs
}
\description{
Returns track IDs for which we have stored history.
}
