% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{load_track_history}
\alias{load_track_history}
\title{Load track history data}
\usage{
load_track_history(track_id, path = "data/tracks")
}
\arguments{
\item{track_id}{Spotify track ID, or "all" to load all tracks}

\item{path}{Directory path containing track CSVs}
}
\value{
tibble of track history
}
\description{
Loads track history from CSV file(s).
}
\examples{
\dontrun{
# Load single track
history <- load_track_history("53iuhJlwXhSER5J2IYYv1W")

# Load all tracks
all_history <- load_track_history("all")
}
}
