% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_chart.R
\name{parse_chart_row}
\alias{parse_chart_row}
\title{Parse a single chart row}
\usage{
parse_chart_row(row)
}
\arguments{
\item{row}{rvest html_element for a table row}
}
\value{
tibble with one row of data
}
\description{
Parse a single chart row
}
\keyword{internal}
