% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_chart.R
\name{parse_chart_table}
\alias{parse_chart_table}
\title{Parse chart HTML table}
\usage{
parse_chart_table(html)
}
\arguments{
\item{html}{rvest html_document}
}
\value{
tibble of chart data
}
\description{
Internal function to extract data from the chart table HTML.
}
\keyword{internal}
