% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_position_change}
\alias{parse_position_change}
\title{Parse position change indicator}
\usage{
parse_position_change(x)
}
\arguments{
\item{x}{Character string with position change}
}
\value{
Named list with \code{type} ("new", "re", "up", "down", "same") and \code{value} (integer)
}
\description{
Parses position change strings like "NEW", "RE", "+46", "-2", "=".
}
\keyword{internal}
